// src/generated/models/Cr22e_proveedoresesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr22e_proveedoresesstatecode = {
  0: 'Activo',
  1: 'Inactivo'
} as const;
export type Cr22e_proveedoresesstatecode = keyof typeof Cr22e_proveedoresesstatecode;
export const Cr22e_proveedoresesstatuscode = {
  1: 'Activo',
  2: 'Inactivo'
} as const;
export type Cr22e_proveedoresesstatuscode = keyof typeof Cr22e_proveedoresesstatuscode;

export interface Cr22e_proveedoresesBase {
  cr22e_activo?: boolean;
  cr22e_correogeneral?: string;
  cr22e_name: string;
  cr22e_observaciones?: string;
  cr22e_proveedoresid: string;
  cr22e_razonsocial?: string;
  cr22e_rfc?: string;
  cr22e_sitioweb?: string;
  cr22e_telefonogeneral?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr22e_proveedoresesstatecode;
  statuscode?: Cr22e_proveedoresesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr22e_proveedoreses extends Cr22e_proveedoresesBase {
  cr22e_activoname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}